Move WordPress IDE to a separate pseudopackage
