+++ src/wordpress-plugin-ide/bundler/babel-plugin-add-import-extension.js
