+++ src/wordpress-plugin-ide/bundler/babel-plugin-transpile-wordpress-imports.ts
