+++ src/wordpress-plugin-ide/bundler/babel-plugin-transpile-wordpress-jsx.ts
