+++ src/wordpress-plugin-ide/bundler/create-amd-loader.ts
