+++ src/wordpress-plugin-ide/bundler/index.ts
