+++ src/wordpress-plugin-ide/bundler/rollup-plugin-commonjs.js
