+++ src/wordpress-plugin-ide/bundler/rollup-plugin-json.ts
