+++ src/wordpress-plugin-ide/bundler/rollup-plugin-provide-dependencies.ts
