+++ src/wordpress-plugin-ide/bundler/rollup-plugin-utils.js
