+++ src/wordpress-plugin-ide/hooks.ts
