+++ src/wordpress-plugin-ide/index.ts
@@ -1,0 +1,2 @@
+export { WordPressPluginIDE } from './WordPressPluginIDE';
+export * from './php-fixtures';
