+++ src/wordpress-plugin-ide/php-fixtures/create-block-plugin.ts
