+++ src/wordpress-plugin-ide/php-fixtures/enable-react-fast-refresh.ts
